#97 fix CPU metrics not working on Windows

+++ server/src/main/java/com/krillsson/sysapi/graphql/QueryResolver.kt
@@ -244,10 +244,10 @@ class QueryResolver : GraphQLQueryResolver {
     }
 
     inner class ProcessorMetricsResolver : GraphQLResolver<CpuLoad> {
-        fun getVoltage(cpuLoad: CpuLoad) = cpuLoad.cpuHealth.voltage
-        fun getFanRpm(cpuLoad: CpuLoad) = cpuLoad.cpuHealth.fanRpm
+        fun getVoltage(cpuLoad: CpuLoad) = cpuLoad.cpuHealth.voltage.toInt()
+        fun getFanRpm(cpuLoad: CpuLoad) = cpuLoad.cpuHealth.fanRpm.toInt()
         fun getFanPercent(cpuLoad: CpuLoad) =
-            cpuLoad.cpuHealth.fanPercent
+            cpuLoad.cpuHealth.fanPercent.toInt()
 
         fun getTemperatures(cpuLoad: CpuLoad) =
             cpuLoad.cpuHealth.temperatures
